refactor(chat): メッセージパーツ描画を専用コンポーネントへ分割
- chat のメッセージパーツ描画を責務ごとに分割し、`MessagePart` の分岐処理を整理
- text パーツの添付ファイル表示ロジックを `text-content` 側へ移動
- thinking と tool-result の描画をそれぞれ専用コンポーネントとして切り出し
- 未実装パーツと未知のパーツ向けに、内容を確認できる共通表示コンポーネントを追加
